% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ZebrafishCAGE}
\alias{ZebrafishCAGE}
\title{CAGE data for zebrafish developmental time course produced by ZEPROME
consortium}
\format{
\code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::RangedSummarizedExperiment}} with genomic
positions of TSSs detected by CAGE as a \code{\link[GenomicRanges:GPos-class]{GenomicRanges::UnstitchedGPos}}
object and frequency of their usage in the 12 developmental stages of
zebrafish as a \code{\link[S4Vectors:DataFrame-class]{S4Vectors::DataFrame}} of \code{\link[S4Vectors:Rle-class]{S4Vectors::Rle}} of \code{integer}s.

The CAGE detected TSSs (mapped to the Zv9 (danRer7) zebrafish genome
assembly) and numbers of CAGE tags were obtained from the original
publication by Nepal \emph{et al.} (Genome Research 2013).
}
\usage{
ZebrafishCAGE
}
\description{
This dataset contains all transcription start sites (TSSs) detected by CAGE
and their frequency of usage in the 12 developmental stages during zebrafish
\emph{Danio rerio} early embryonic development.
}
\examples{

ZebrafishDevelopmentalCAGE::ZebrafishCAGE
ZebrafishDevelopmentalCAGE::ZebrafishCAGE |> rowRanges()
ZebrafishDevelopmentalCAGE::ZebrafishCAGE |> assay()
ZebrafishDevelopmentalCAGE::ZebrafishCAGE |> colData()
}
\references{
Nepal \emph{et al.} (2013). \emph{Dynamic regulation of coding and non-coding
transcription initiation landscape at single nucleotide resolution during
vertebrate embryogenesis}. Genome Research \strong{23}:1938-1950.
}
\keyword{datasets}
