% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ZebrafishSamples}
\alias{ZebrafishSamples}
\title{List of CAGE samples available in this package}
\format{
A \code{data.frame} with 3 columns specifying the name of the dataset
(only one dataset named \code{ZebrafishCAGE}), the name of the group (only one
group named \code{development}) and the names of individual CAGE samples,
respectively.  These labels should be used in the \code{\link[CAGEr:importPublicData]{CAGEr::importPublicData}}
to import selected CAGE samples into a \code{\link[CAGEr:CAGEexp-class]{CAGEr::CAGEexp}} object.
}
\usage{
ZebrafishSamples
}
\description{
This datasets lists all CAGE samples for zebrafish developmental timecourse
produced by ZEPROME consortium (Nepal \emph{et al.} Genome Research 2013) that are
available in this package.
}
\examples{
ZebrafishDevelopmentalCAGE::ZebrafishSamples
}
\references{
Nepal \emph{et al.} (2013). \emph{Dynamic regulation of coding and non-coding
transcription initiation landscape at single nucleotide resolution during
vertebrate embryogenesis}. Genome Research \strong{23}:1938-1950.
}
\keyword{datasets}
